refactor: Formatting -> MemoryFormatter
